refactor(tui): rename chat-ui to chat-app
